Moved parser.py from the ipython module to the util module
